Integrate Upload and Create Spreadsheet + Bug fixes
Integration of Create and Upload Section caused issues like add rows and add columns functionality not working, data leakage from Upload Section to Create Section. Solved these issues.
