+++ package.json
@@ -73,7 +73,6 @@
 		"qr-code-styling": "^1.6.0-rc.1",
 		"svelte-datepicker": "^1.0.0",
 		"svelte-tags-input": "^5.0.0",
-		"tabulator-tables": "^5.5.2",
 		"xlsx": "^0.18.5"
 	}
 }
